refactor(babel-config): gave the config writer a noun-based name
