+++ src/dialects/babel/config/index.js
@@ -1,3 +1,3 @@
-export {default as write} from './write.js';
+export {default as write} from './writer.js';
 export {default as load} from './loader.js';
 export {default as addIgnore} from './ignore-adder.js';
